separate extension class / add obsolete attribute to some classes
