+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/MediaPipeTracker/Tasks/HolisticLandmarkerResultExtensions.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8ab8c463aa5afa3428968cc093608d33
